added OnBlockAdd
