+++ code/go/0chain.net/chaincore/chain/entity.go
@@ -199,6 +199,8 @@ type Chain struct {
 
 	// compute state
 	computeBlockStateC chan struct{}
+
+	OnBlockAdded func(b *block.Block)
 }
 
 // SyncBlockReq represents a request to sync blocks, it will be
@@ -413,6 +415,7 @@ func NewChainFromConfig() *Chain {
 
 	chain.NotarizedBlocksCounts = make([]int64, chain.MinGenerators()+1)
 	client.SetClientSignatureScheme(chain.ClientSignatureScheme())
+
 	return chain
 }
 
@@ -486,6 +489,8 @@ func (c *Chain) Initialize() {
 	c.minersStake = make(map[datastore.Key]int)
 	c.magicBlockStartingRounds = make(map[int64]*block.Block)
 	c.MagicBlockStorage = round.NewRoundStartingStorage()
+	c.OnBlockAdded = func(b *block.Block) {
+	}
 }
 
 /*SetupEntity - setup the entity */
@@ -694,8 +699,6 @@ func (c *Chain) addBlock(b *block.Block) *block.Block {
 	}
 	c.blocks[b.Hash] = b
 
-	c.pushBlockEvent(b)
-
 	if b.PrevBlock == nil {
 		if pb, ok := c.blocks[b.PrevHash]; ok {
 			b.SetPreviousBlock(pb)
@@ -708,20 +711,9 @@ func (c *Chain) addBlock(b *block.Block) *block.Block {
 			break
 		}
 	}
-	return b
-}
 
-func (c *Chain) pushBlockEvent(b *block.Block) {
-	err, ev := block.CreateBlockEvent(b)
-	if err != nil {
-		logging.Logger.Error("emit block event error", zap.Error(err))
-	}
-	go func() {
-		rootContext := common.GetRootContext()
-		ctx, cancel := context.WithTimeout(rootContext, 5*time.Second)
-		defer cancel()
-		c.GetEventDb().AddEvents(ctx, []event.Event{ev})
-	}()
+	c.OnBlockAdded(b)
+	return b
 }
 
 /*GetBlock - returns a known block for a given hash from the cache */
